Extracted connection code
