+++ src/connect.rs
@@ -1,0 +1,231 @@
+//! Node connection utilities.
+
+use crate::{node_label::InternedNodeLabel, AudioContext};
+use crate::{MainBus, Node, NodeLabel};
+use bevy_ecs::prelude::*;
+use bevy_log::{error_once, warn_once};
+use bevy_utils::HashMap;
+use firewheel::node::NodeID;
+
+/// A target for node connections.
+///
+/// [`ConnectTarget`] can be constructed manually or
+/// used as a part of the [`ConnectNode`] API.
+#[derive(Debug, Clone)]
+pub enum ConnectTarget {
+    /// A global label such as [`MainBus`].
+    Label(InternedNodeLabel),
+    /// An audio entity.
+    Entity(Entity),
+    /// An existing node from the audio graph.
+    Node(NodeID),
+}
+
+/// A pending connection between two nodes.
+#[derive(Debug, Clone)]
+pub struct PendingConnection {
+    pub target: ConnectTarget,
+    /// An optional [`firewheel`] port mapping.
+    ///
+    /// The first tuple element represents the source output,
+    /// and the second tuple element represents the sink input.
+    ///
+    /// If an explicit port mapping is not provided,
+    /// `[(0, 0), (1, 1)]` is used.
+    pub ports: Option<Vec<(u32, u32)>>,
+}
+
+impl From<NodeID> for ConnectTarget {
+    fn from(value: NodeID) -> Self {
+        Self::Node(value)
+    }
+}
+
+impl<T> From<T> for ConnectTarget
+where
+    T: NodeLabel,
+{
+    fn from(value: T) -> Self {
+        Self::Label(value.intern())
+    }
+}
+
+impl From<Entity> for ConnectTarget {
+    fn from(value: Entity) -> Self {
+        Self::Entity(value)
+    }
+}
+
+/// The set of all [`PendingConnection`]s for an entity.
+///
+/// These connections are drained and synchronized with the
+/// audio graph in the [SeedlingSystems::Connect] set.
+#[derive(Debug, Default, Component)]
+pub struct PendingConnections(Vec<PendingConnection>);
+
+impl PendingConnections {
+    /// Push a new pending connection.
+    pub fn push(&mut self, connection: PendingConnection) {
+        self.0.push(connection)
+    }
+}
+
+/// An [`EntityCommands`] extension trait for connecting node entities.
+///
+/// These methods provide only source -> sink connections. The source
+/// is the receiver and the sink is the provided target.
+///
+/// [`EntityCommands`]: bevy_ecs::prelude::EntityCommands
+pub trait ConnectNode {
+    /// Queue a connection from this entity to the target.
+    ///
+    /// ```
+    /// # use bevy::prelude::*;
+    /// # use bevy_seedling::{MainBus, VolumeNode, ConnectNode};
+    /// # fn system(mut commands: Commands) {
+    /// // Connect a node to the MainBus.
+    /// let node = commands.spawn(VolumeNode { normalized_volume: 0.5 }).connect(MainBus).id();
+    ///
+    /// // Connect another node to the one we just spawned.
+    /// commands.spawn(VolumeNode { normalized_volume: 0.25 }).connect(node);
+    /// # }
+    /// ```
+    ///
+    /// By default, this provides a port connection of `[(0, 0), (1, 1)]`,
+    /// which represents a simple stereo connection.
+    /// To provide a specific port mapping, use [`connect_with`][ConnectNode::connect_with].
+    ///
+    /// The connection is deferred, finalizing in the [`SeedlingSystems::Connect`] set.
+    fn connect(&mut self, target: impl Into<ConnectTarget>) -> &mut Self;
+
+    /// Queue a connection from this entity to the target with the provided port mappings.
+    ///
+    /// The connection is deferred, finalizing in the [`SeedlingSystems::Connect`] set.
+    fn connect_with(&mut self, target: impl Into<ConnectTarget>, ports: &[(u32, u32)])
+        -> &mut Self;
+}
+
+impl ConnectNode for EntityCommands<'_> {
+    fn connect(&mut self, target: impl Into<ConnectTarget>) -> &mut Self {
+        let target = target.into();
+
+        self.entry::<PendingConnections>()
+            .or_default()
+            .and_modify(|mut pending| {
+                pending.push(PendingConnection {
+                    target,
+                    ports: None,
+                });
+            });
+
+        self
+    }
+
+    fn connect_with(
+        &mut self,
+        target: impl Into<ConnectTarget>,
+        ports: &[(u32, u32)],
+    ) -> &mut Self {
+        let target = target.into();
+        let ports = ports.to_vec();
+
+        self.entry::<PendingConnections>()
+            .or_default()
+            .and_modify(|mut pending| {
+                pending.push(PendingConnection {
+                    target,
+                    ports: Some(ports),
+                });
+            });
+
+        self
+    }
+}
+
+const DEFAULT_CONNECTION: &[(u32, u32)] = &[(0, 0), (1, 1)];
+
+// this has turned into a bit of a monster
+pub(crate) fn process_connections(
+    mut connections: Query<(&mut PendingConnections, &Node)>,
+    targets: Query<&Node>,
+    node_map: Res<NodeMap>,
+    mut context: ResMut<AudioContext>,
+) {
+    context.with(|context| {
+        for (mut pending, source_node) in connections.iter_mut() {
+            pending.0.retain(|connection| {
+                let ports = connection.ports.as_deref().unwrap_or(DEFAULT_CONNECTION);
+
+                let target_entity = match connection.target {
+                    ConnectTarget::Entity(entity) => entity,
+                    ConnectTarget::Label(label) => {
+                        let Some(entity) = node_map.get(&label) else {
+                            warn_once!("tried to connect audio node to label with no node");
+
+                            // We may need to wait for the intended label to be spawned.
+                            return true;
+                        };
+
+                        *entity
+                    }
+                    ConnectTarget::Node(dest_node) => {
+                        // no questions asked, simply connect
+                        if let Err(e) = context.connect(source_node.0, dest_node, ports, false) {
+                            error_once!("failed to connect audio node to target: {e}");
+                        }
+
+                        // if this fails, the target node must have been removed from the graph
+                        return false;
+                    }
+                };
+
+                let target = match targets.get(target_entity) {
+                    Ok(t) => t,
+                    Err(_) => {
+                        error_once!("failed to fetch audio node entity {target_entity:?}");
+                        return false;
+                    }
+                };
+
+                if let Err(e) = context.connect(source_node.0, target.0, ports, false) {
+                    error_once!("failed to connect audio node to target: {e}");
+                }
+
+                false
+            });
+        }
+    });
+}
+
+/// A map that associates [`NodeLabel`]s with audio
+/// graph nodes.
+///
+/// This will be automatically synchronized for
+/// entities with both a [`Node`] and [`NodeLabel`].
+#[derive(Default, Debug, Resource)]
+pub struct NodeMap(HashMap<InternedNodeLabel, Entity>);
+
+impl core::ops::Deref for NodeMap {
+    type Target = HashMap<InternedNodeLabel, Entity>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl core::ops::DerefMut for NodeMap {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+/// Automatically connect nodes without manual connections to the main bus.
+pub(crate) fn auto_connect(
+    nodes: Query<Entity, (With<Node>, Without<PendingConnections>)>,
+    mut commands: Commands,
+) {
+    for node in nodes.iter() {
+        commands.entity(node).connect(MainBus);
+    }
+}
+
